+++ darwin/Classes/ClippingAudioSource.m
@@ -14,8 +14,8 @@
     self = [super initWithId:sid];
     NSAssert(self, @"super init cannot be nil");
     _audioSource = audioSource;
-    _start = start == [NSNull null] ? kCMTimeZero : CMTimeMake([start intValue], 1000);
-    _end = end == [NSNull null] ? kCMTimeInvalid : CMTimeMake([end intValue], 1000);
+    _start = start == [NSNull null] ? kCMTimeZero : CMTimeMake([start longLongValue], 1000000);
+    _end = end == [NSNull null] ? kCMTimeInvalid : CMTimeMake([end longLongValue], 1000000);
     return self;
 }
 
@@ -56,6 +56,8 @@
     if (!completionHandler || (self.playerItem.status == AVPlayerItemStatusReadyToPlay)) {
         CMTime absPosition = CMTimeAdd(_start, position);
         [_audioSource.playerItem seekToTime:absPosition toleranceBefore:kCMTimeZero toleranceAfter:kCMTimeZero completionHandler:completionHandler];
+    } else {
+        [super seek:position completionHandler:completionHandler];
     }
 }