+++ darwin/Classes/UriAudioSource.m
@@ -14,7 +14,7 @@
     NSAssert(self, @"super init cannot be nil");
     _uri = uri;
     if ([_uri hasPrefix:@"file://"]) {
-        _playerItem = [[IndexedPlayerItem alloc] initWithURL:[NSURL fileURLWithPath:[_uri substringFromIndex:7]]];
+        _playerItem = [[IndexedPlayerItem alloc] initWithURL:[NSURL fileURLWithPath:[[_uri stringByReplacingPercentEscapesUsingEncoding:NSUTF8StringEncoding] substringFromIndex:7]]];
     } else {
         _playerItem = [[IndexedPlayerItem alloc] initWithURL:[NSURL URLWithString:_uri]];
     }
@@ -56,6 +56,8 @@
             position = CMTimeAdd(position, range.start);
         }
         [_playerItem seekToTime:position toleranceBefore:kCMTimeZero toleranceAfter:kCMTimeZero completionHandler:completionHandler];
+    } else {
+        [super seek:position completionHandler:completionHandler];
     }
 }